Update for cluster migration
As part of a cluster migration, we needed to change a number of things.

The fedora build was removing python2 as part of our cleanup process.
Commenting out the uninstall line in the Dockerfile for now - in the future
the build will be slightly more efficient if we determine what python2 depends
on and clean up the other things.

We were using gluster for our persistent storage previously.  That is no longer
an option.  We've switched to cephRBD, which doesn't support ReadWriteMany.  So
the shiftzilla website will now be placed on the bugzilla-state pvc which is
ReadWriteMany. The database will remain on the cephRBD shiftzilla-state pvc,
which will only be accessed by a single pod at a time (build, cleanup, backup).